refactor(viewer): extract comment/context-menu + map pointer interactions
Continues breaking up the two God Components. Two independent extractions in
one PR (both faithful, verbatim moves with only dependencies rewired):

ViewerStage (1165 -> 920):
- useCommentFlow — the create/edit popover lifecycle, comment mode + side panel,
  and seek-to-comment helpers (glue between useReplay and useComments).
- useContextMenu — the right-click target state + actions (edit/delete/resize a
  comment, add comment, follow player, copy setpos); comment edit/create delegate
  back to the comment flow, follow is injected from the stage.

ViewerMap (2369 -> 1723):
- useMapInteractions — all pointer handling (wheel/pan, comment placement/hover/
  area-resize, coach board: draw shapes, place/drag grenades, move/rotate players)
  plus the hit-testing and the override/drag-aware pose resolution. The render
  loop stays in the component and reads the transient drag/hover state back
  through a getter object, calls the interaction overlays (rotate handle, area
  handles, hover ring) and `coachPosOf`, and shares `commentHitboxes` via a ref.

type-check and production build pass. Canvas/keyboard behavior isn't covered by
type-check, so this needs a manual pass (see PR notes).
